requirements.txt
async-cache
dotenv
databases
discord.py